requirements.txt
Django==1.8.3
django-debug-toolbar==1.3.2
celery==3.1.8
django-celery==3.1.16
Pillow==3.0.0
django-multiselectfield==0.1.3
django-solo==1.1.1
